requirements.txt
torch
torchvision
transformers
openai
requests
streamlit
Pillow
dotenv